refactor: update file and folder names to lowercase
Standardized file and folder names to lowercase for consistency. Updated all associated imports and exports to match the new naming convention.
