+++ apps/nextjs/paths.txt
@@ -1,25 +1,12 @@
-src/components/pages/home/__tests__/Container.integration.test.tsx
-src/components/pages/home/Component.tsx
-src/components/pages/home/components/Content/Content.tsx
-src/components/pages/home/components/Content/hooks/index.ts
-src/components/pages/home/components/Content/hooks/useHomeContentDom.ts
-src/components/pages/home/components/Content/hooks/useHomeContentVerticalScroll.ts
-src/components/pages/home/components/Content/index.ts
-src/components/pages/home/components/FavoriteProjects/FavoriteProjects.tsx
-src/components/pages/home/components/FavoriteProjects/index.ts
-src/components/pages/home/components/Header/Header.tsx
-src/components/pages/home/components/Header/index.ts
-src/components/pages/home/components/ProjectsContainer/index.ts
-src/components/pages/home/components/ProjectsContainer/ProjectsContainer.tsx
-src/components/pages/home/components/RecentProjects/index.ts
-src/components/pages/home/components/RecentProjects/RecentProjects.tsx
-src/components/pages/home/components/SkeletonHome.tsx
-src/components/pages/home/components/TasksDueSoon/index.ts
-src/components/pages/home/components/TasksDueSoon/ListItem.tsx
-src/components/pages/home/components/TasksDueSoon/TasksDueSoon.tsx
-src/components/pages/home/components/TasksDueSoon/types.ts
-src/components/pages/home/Container.tsx
-src/components/pages/home/Home.stories.tsx
-src/components/pages/home/hooks/index.ts
-src/components/pages/home/hooks/useHomeTaskDetail.tsx
-src/components/pages/home/index.ts
+src/components/pages/inbox/Component.tsx
+src/components/pages/inbox/components/Activity/Activity.tsx
+src/components/pages/inbox/components/Activity/index.ts
+src/components/pages/inbox/components/Archive/Archive.tsx
+src/components/pages/inbox/components/Archive/index.ts
+src/components/pages/inbox/components/Header/Header.tsx
+src/components/pages/inbox/components/Header/index.ts
+src/components/pages/inbox/components/Header/Tabs.tsx
+src/components/pages/inbox/Container.tsx
+src/components/pages/inbox/Inbox.stories.tsx
+src/components/pages/inbox/index.ts
+src/components/pages/inbox/providers/Provider.tsx
