--- apps/nextjs/src/components/pages/inbox/components/Activity/index.ts
@@ -1,1 +1,0 @@
-export { Activity } from './Activity';
