--- apps/nextjs/src/components/pages/inbox/components/Archive/index.ts
@@ -1,1 +1,0 @@
-export { Archive } from './Archive';
