+++ apps/nextjs/src/components/pages/inbox/components/activity/index.ts
@@ -1,0 +1,1 @@
+export { Activity } from './activity';
