+++ apps/nextjs/src/components/pages/inbox/components/archive/index.ts
@@ -1,0 +1,1 @@
+export { Archive } from './archive';
