+++ apps/nextjs/src/components/pages/inbox/components/header/header.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react';
 import { Flex } from '@/components/ui/flex';
-import { Tabs } from './Tabs';
+import { Tabs } from './tabs';
 
 export const Header = memo(function Header() {
   return (
