+++ apps/nextjs/src/components/pages/inbox/components/header/index.ts
@@ -1,0 +1,1 @@
+export { Header } from './header';
